pyproject.toml
[tool.poetry]
name = "prostoprosport_news_fetcher"
version = "0.1.0"
description = "Script to fetch news using API and convert to wiki-text"
authors = ["Artiom Khandamirov <t9max@yandex.ru>"]
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.8,<=3.11"
beautifulsoup4 = "^4.11.1"
click = "^8.1.3"
aiohttp = "^3.8.1"
tortoise-orm = {extras = ["accel", "asyncpg"], version = "^0.19.1"}
feedparser = "^6.0.10"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
flake8-bandit = "^3.0.0"
flake8-deprecated = "^1.3"
autopep8 = "^1.6.0"
rope = "^1.2.0"
mypy = "^0.961"
flake8-isort = "^4.1.1"
flake8-bugbear = "^22.6.22"
flake8-builtins = "^1.5.3"
flake8-eradicate = "^1.2.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
